backend/requirements.txt
fastapi
uvicorn
python-dotenv
msal
requests
openai
langchain
qdrant-client
pymupdf
python-docx